+++ src/DataPersister/UserBookingDataPersister.php
@@ -4,12 +4,15 @@ namespace App\DataPersister;
 
 use ApiPlatform\Core\DataPersister\ContextAwareDataPersisterInterface;
 use App\Entity\Main\UserBooking;
+use App\Enum\NotificationTypeEnum;
 use App\Exception\MicrosoftGraphCommunicationException;
 use App\Exception\UserBookingException;
+use App\Message\SendUserBookingNotificationMessage;
 use App\Security\Voter\UserBookingVoter;
 use App\Service\BookingServiceInterface;
 use Symfony\Component\HttpKernel\Exception\AccessDeniedHttpException;
 use Symfony\Component\HttpKernel\Exception\HttpException;
+use Symfony\Component\Messenger\MessageBusInterface;
 use Symfony\Component\Security\Core\Security;
 
 class UserBookingDataPersister implements ContextAwareDataPersisterInterface
@@ -17,6 +20,7 @@ class UserBookingDataPersister implements ContextAwareDataPersisterInterface
     public function __construct(
         private readonly BookingServiceInterface $bookingService,
         private readonly Security $security,
+        private readonly MessageBusInterface $bus,
     ) {
     }
 
@@ -34,6 +38,11 @@ class UserBookingDataPersister implements ContextAwareDataPersisterInterface
                 }
 
                 $this->bookingService->deleteBooking($data);
+
+                $this->bus->dispatch(new SendUserBookingNotificationMessage(
+                    $data,
+                    NotificationTypeEnum::DELETE_SUCCESS
+                ));
             }
         } catch (MicrosoftGraphCommunicationException|UserBookingException $e) {
             throw new HttpException($e->getCode(), 'Booking could not be deleted.');
@@ -49,6 +58,11 @@ class UserBookingDataPersister implements ContextAwareDataPersisterInterface
                 }
 
                 $this->bookingService->updateBooking($data);
+
+                $this->bus->dispatch(new SendUserBookingNotificationMessage(
+                    $data,
+                    NotificationTypeEnum::UPDATE_SUCCESS
+                ));
             }
         } catch (MicrosoftGraphCommunicationException|UserBookingException $e) {
             throw new HttpException($e->getCode(), 'Booking could not be updated.');