+++ CHANGELOG.md
@@ -9,4 +9,19 @@ See [keep a changelog](https://keepachangelog.com/en/1.0.0/) for information abo
 
 ## [Unreleased]
 
-- TODO: Start changelog after first release.
+### Added
+
+- [PR-89](https://github.com/itk-dev/book_aarhus/pull/89): Added check to avoid getting many user-bookings in the past.
+- [PR-89](https://github.com/itk-dev/book_aarhus/pull/89): Added notifications to user on edit and delete.
+- [PR-89](https://github.com/itk-dev/book_aarhus/pull/89): Added notifications to resource on edit and delete.
+
+### Changed
+
+- [PR-89](https://github.com/itk-dev/book_aarhus/pull/89): Fixed issue with GET /user-bookings where pagination
+was not handled correctly.
+- [PR-90](https://github.com/itk-dev/book_aarhus/pull/90): Updated fixtures to match expected data for resource
+categories.
+
+## [1.0.0] - YYYY-MM-DD
+
+- First release.